+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Next release
 
 - chore: add headers to da calldata, fix eth da in sovereign mode
+- chore: feature flags for avail and celestia DA
+- feat(rpc): added support for v0.5.1 JSON-RPC specs
+- feat(rpc): added ordered messages/events in trace fields
+- feat(rpc): support for starknet.rs v0.5.1 version
+- feat(rpc): added execution resources in trace fields
+- feat(rpc): added state diff field in trace fields
 - refactor: removed benchmarking folder and traces of CI pipeline
 - fix: decouple is_query into is_query and offset_version
 - feat: add sierra to casm class hash mapping to genesis assets